% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_date_to_bandname.R
\name{add_date_to_bandname.ee.image.Image}
\alias{add_date_to_bandname.ee.image.Image}
\title{add_date_to_band_name}
\usage{
\method{add_date_to_bandname}{ee.image.Image}(x)
}
\value{
a date to band name in image.
}
\description{
add_date_to_band_name
}
