% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_date_to_bandname.R
\name{add_date_to_bandname.ee.imagecollection.ImageCollection}
\alias{add_date_to_bandname.ee.imagecollection.ImageCollection}
\title{add_date_to_band_name}
\usage{
\method{add_date_to_bandname}{ee.imagecollection.ImageCollection}(x)
}
\description{
add_date_to_band_name
}
