% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_date_to_bandname.R
\name{add_date_to_bandname}
\alias{add_date_to_bandname}
\title{add_date_to_band_name}
\usage{
add_date_to_bandname(x)
}
\arguments{
\item{x}{ee$ImageCollection or ee$Image}
}
\value{
a date to band name in x.
}
\description{
append date to band name
}
