% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ee.R
\name{as_ee}
\alias{as_ee}
\title{as_ee tidyee to ee$ImageCollection or ee$Image}
\usage{
as_ee(x)
}
\arguments{
\item{x}{tidyee}
}
\value{
ee$ImageCollection or ee$Image
}
\description{
as_ee tidyee to ee$ImageCollection or ee$Image
}
\examples{
\dontrun{
library(rgee)
library(tidyrgee)

modis_ic <- ee$ImageCollection("MODIS/006/MOD13Q1")

# create tidyee class
modis_ic_tidy <- as_tidyee(modis_ic)
# convert back to origina ee$ImageCollection class
modis_ic_tidy |>
  as_ee()
}
}
