% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind_ics}
\alias{bind_ics}
\title{bind ImageCollections}
\usage{
bind_ics(x)
}
\arguments{
\item{x}{list of tidyee objects}
}
\value{
tidyee object containing single image collection and vrt
}
\description{
bind ImageCollections
}
\examples{
\dontrun{
library(tidyrgee)
library(rgee)
ee_Initialize()
modis_ic <- ee$ImageCollection("MODIS/006/MOD13Q1")
modis_ic_tidy <- as_tidyee(modis_ic)
modis_tidy_list <- modis_tidy |>
group_split(month)
modis_tidy_list |>
  bind_ics()
}

}
