% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tidyee.R
\name{create_tidyee}
\alias{create_tidyee}
\title{create_tidyee}
\usage{
create_tidyee(x, vrt)
}
\arguments{
\item{x}{ee$ImageCollection}

\item{vrt}{virtual table}
}
\value{
tidyee class list object
}
\description{
helper function to assign new tidyee when running \code{as_tidyee}
}
