% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_range_imageCol}
\alias{date_range_imageCol}
\title{date_range_imageCol}
\usage{
date_range_imageCol(x)
}
\arguments{
\item{x}{imageCollection or image}
}
\value{
sorted date vector (length 2)
}
\description{
a fast-working helper function to extract min and max date ranges for image collections
}
\examples{
\donrun{
library(tidyrgee)
ee_Initialize()
modis_ic <- ee$ImageCollection("MODIS/006/MOD13Q1")
date_range_imageCol(modis_ic)
}

}
