% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_temporal_composites.R
\name{ee_composite}
\alias{ee_composite}
\title{ee_composite}
\usage{
ee_composite(x, stat, ...)
}
\arguments{
\item{x}{tidyee object containing \code{ee$ImageCollection}}

\item{stat}{A \code{character} indicating what to reduce the imageCollection by,
e.g. 'median' (default), 'mean',  'max', 'min', 'sum', 'sd', 'first'.}

\item{...}{other arguments}
}
\value{

}
\description{
ee_composite
}
