% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_temporal_filters.R
\name{ee_month_filter}
\alias{ee_month_filter}
\title{ee_month_filter}
\usage{
ee_month_filter(imageCol, month, ...)
}
\arguments{
\item{imageCol}{ee$ImageCollection}

\item{month}{\code{numeric} vector containing month values (1-12)}

\item{...}{other arguments}
}
\value{
ee$ImageCollection or ee$Image filtered by month
}
\description{
ee_month_filter
}
