% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_roll.R
\name{ee_roll}
\alias{ee_roll}
\title{ee_roll}
\usage{
ee_roll(x, window, stat)
}
\arguments{
\item{x}{imageCollection}

\item{window}{\code{numeric} look back in days}

\item{stat}{\code{character} stat/function to roll with}
}
\value{
imageCollection where each image represents the rolling sum of for specified window "right" aligned
}
\description{
rolling statistics on imageCollection. Written to be analagous to \code{terra::roll}. Currently only has options
for "right" aligned rolling statistcs and window units of days.
}
\section{TODO}{
 a.) include window_unit argument, b.) include alignment arg, c.) write tidyee/ic methods
}

\examples{
\dontrun{
library(rgee)
library(tidyverse)
 ee_Initialize()
 chirps_link <- "UCSB-CHG/CHIRPS/DAILY"
 chirps <- ee$ImageCollection(chirps_link)
 x <- chirps$filterDate("2021-01-01","2022-01-01")
 rolling_10_max <- ee_roll_stat(x = x,window = 10, stat="sum")
}
}
