% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_temporal_composites.R
\name{ee_year_composite}
\alias{ee_year_composite}
\title{ee_year_composite}
\usage{
ee_year_composite(imageCol, stat, start_date, end_date, ...)
}
\arguments{
\item{imageCol}{ee$ImageCollection}

\item{stat}{A \code{character} indicating what to reduce the imageCollection by,
e.g. 'median' (default), 'mean',  'max', 'min', 'sum', 'sd', 'first'.}

\item{start_date}{\code{character} format date, e.g. "2018-10-23".}

\item{end_date}{\code{character} format date, e.g. "2018-10-23".}

\item{...}{other arguments}
}
\description{
ee_year_composite
}
