% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_temporal_composites.R
\name{ee_year_month_composite}
\alias{ee_year_month_composite}
\title{Pixel-level composite by year and month}
\usage{
ee_year_month_composite(x, stat, ...)
}
\arguments{
\item{stat}{A \code{character} indicating what to reduce the imageCollection by,
e.g. 'median' (default), 'mean',  'max', 'min', 'sum', 'sd', 'first'.}

\item{startDate}{\code{character} format date, e.g. "2018-10-23".}

\item{endDate}{\code{character} format date, e.g. "2018-10-23".}

\item{months}{\code{numeric} vector, e.g. c(1,12).}
}
\description{
Pixel-level composite by year and month
}
