% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_temporal_filters.R
\name{ee_year_month_filter}
\alias{ee_year_month_filter}
\title{ee_year_month_filter}
\usage{
ee_year_month_filter(imageCol, year, month, ...)
}
\arguments{
\item{imageCol}{ee$ImageCollection}

\item{year}{\code{numeric} vector contain years to filter}

\item{month}{\code{numeric} vector contain months to filter}

\item{...}{other arguments}
}
\value{
ee$ImageCollection or ee$Image filtered by year & month
}
\description{
ee_year_month_filter
}
