% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_ic}
\alias{filter_ic}
\title{Filter ImageCollection}
\usage{
filter_ic(x, info)
}
\arguments{
\item{setup_list}{A list with params.}
}
\description{
This filters the ImageCollection based on the args passed from the user
}
