% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geometry_type_is_unique}
\alias{geometry_type_is_unique}
\title{geometry_type_is_unique}
\usage{
geometry_type_is_unique(x)
}
\arguments{
\item{x}{sf object}
}
\value{
\code{logical} indicating whether geometry type is unique in sf object
}
\description{
geometry_type_is_unique
}
