% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_by.ee.imagecollection.ImageCollection}
\alias{group_by.ee.imagecollection.ImageCollection}
\title{Title}
\usage{
\method{group_by}{ee.imagecollection.ImageCollection}(x, ...)
}
\arguments{
\item{...}{group_by variables  (must be "month", "year" or combination of the two)}
}
\value{
imageCollection with grouped_vars atttribute
}
\description{
Title
}
\examples{
\dontrun{
library(tidyrgee)
ee_Initialize()
modis_ic <- ee$ImageCollection("MODIS/006/MOD13Q1")
modis_ic |>
   filter(date>="2016-01-01",date<="2019-12-31") |>
   group_by(year)
}

}
