% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_by}
\alias{group_by}
\title{group_by}
\usage{
group_by(x, ...)
}
\arguments{
\item{x}{ee$ImageCollection}

\item{...}{group_by variables  (must be "month", "year" or combination of the two)}
}
\value{
ee$ImageCollection with grouped_vars attribute
}
\description{
group_by
}
\examples{
\dontrun{
library(tidyrgee)
ee_Initialize()
modis_ic <- ee$ImageCollection("MODIS/006/MOD13Q1")
modis_ic |>
   filter(date>="2016-01-01",date<="2019-12-31") |>
   group_by(year)
}
}
