% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_join.R
\name{inner_join}
\alias{inner_join}
\title{ic_join_bands}
\usage{
inner_join(x, y, by)
}
\arguments{
\item{x, y}{A pair of tidyee objects containing ee$ImageCollections}

\item{by}{A character vector of variables to join by.}
}
\value{
An object of the same type as \code{x}. The output has the following properties:
Same number of images as \code{x}
Total number of bands equal the number of bands in \code{x} plus the number of bands in \code{y}
}
\description{
ic_join_bands
}
