% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last_day_of_month}
\alias{last_day_of_month}
\title{last_day_of_month}
\usage{
last_day_of_month(year, month_numeric)
}
\arguments{
\item{month_numeric}{\code{numeric} vector containing months of interest}
}
\value{
\code{numeric} vector which the last day of each month
}
\description{
last_day_of_month
}
