% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate}
\alias{mutate}
\title{mutate columns into tidyee vrt which can later be used to modify tidyee ImageCollection}
\arguments{
\item{x}{tidyee class object (list of ee_ob, vrt)}

\item{...}{mutate arguments}
}
\value{
tidyee object with vrt component containing new mutated cols
}
\description{
mutate columns into tidyee vrt which can later be used to modify tidyee ImageCollection
}
\examples{
\dontrun{
library(tidyrgee)
library(rgee)
ee_Initialize()
modis_ic <- ee$ImageCollection("MODIS/006/MOD13Q1")
modis_ic_tidy <- as_tidyee(modis_ic)
}
}
\seealso{
\code{\link[dplyr]{mutate}} for information about mutate on normal data tables.
}
