% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tidyee}
\alias{print.tidyee}
\title{print tidyee}
\usage{
\method{print}{tidyee}(x, ...)
}
\arguments{
\item{x}{tidyee object}

\item{...}{additional arguments}
}
\value{
printed tidyee object
}
\description{
print tidyee
}
