% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_stdDev_bands}
\alias{rename_stdDev_bands}
\title{rename_stdDev_bands}
\usage{
rename_stdDev_bands(x)
}
\value{
x ee$Image/ImageCollection with \verb{.*_stdDev$} bands renamed to \verb{.*_sd$}
}
\description{
rename_stdDev_bands
}
