% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rstat_to_eestat}
\alias{rstat_to_eestat}
\title{rstat_to_eestat - helper function - useful in ee_*_composite functions to get bandNames from vrt}
\usage{
rstat_to_eestat(fun)
}
\arguments{
\item{fun}{reducer/statistic using typical r-syntax (character)}
}
\description{
rstat_to_eestat - helper function - useful in ee_*_composite functions to get bandNames from vrt
}
