% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_idx.R
\name{set_idx}
\alias{set_idx}
\title{set_idx}
\usage{
set_idx(x, idx_name = "tidyee_index")
}
\arguments{
\item{x}{tidyee or \code{ee$ImageCollection} class object}

\item{idx_name}{name for index to create (default = "tidyee_index")}
}
\value{
tidyee or \code{ee$ImageCollection} class object with new index containing sequential 0-based indexing
}
\description{
set_idx
}
\examples{
\dontrun{
library(rgee)
library(tidyrgee)
ee_Initialize()
modis_link <- "MODIS/006/MOD13Q1"
modisIC <- ee$ImageCollection(modis_link)
modis_ndvi_tidy <- as_tidyee(modisIC) |>
  select("NDVI")
modis_ndvi_tidy |>

}
}
