% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stat_to_reducer_full}
\alias{stat_to_reducer_full}
\title{stat_to_reducer_full - helper function}
\usage{
stat_to_reducer_full(fun)
}
\arguments{
\item{fun}{reducer/statistic using typical r-syntax (character)}
}
\description{
stat_to_reducer_full - helper function
}
