% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str}
\alias{str}
\alias{str.ee.imagecollection.ImageCollection}
\alias{str.ee.image.Image}
\title{str}
\usage{
\method{str}{ee.imagecollection.ImageCollection}(object, ...)

\method{str}{ee.image.Image}(object, ...)
}
\value{
A blank str method

A blank str method
}
\description{
str

str
}
