% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_pixels}
\alias{summarise_pixels}
\title{Summary pixel-level stats for ee$ImageCollection or tidyrgee objects with ImageCollections}
\usage{
summarise_pixels(x, stat, ...)
}
\arguments{
\item{x}{ee$Image or ee$ImageCollection}

\item{stat}{stat/function to apply}

\item{...}{other arguments}
}
\description{
Summary pixel-level stats for ee$ImageCollection or tidyrgee objects with ImageCollections
}
