% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungroup.R
\name{ungroup}
\alias{ungroup}
\title{ungroup}
\arguments{
\item{x}{tidyee object}

\item{...}{ungroup args}
}
\description{
ungroup
}
\seealso{
\code{\link[dplyr]{ungroup}} for information about ungroup on normal data tables.
}
