% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vrt_band_names}
\alias{vrt_band_names}
\title{vrt_band_names}
\usage{
vrt_band_names(x)
}
\arguments{
\item{x}{tidyee class object}
}
\value{
a character vector of band_names
}
\description{
vrt_band_names
}
