% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_temporal_composites.R
\name{ee_composite}
\alias{ee_composite}
\alias{ee_composite.tidyee}
\title{ee_composite}
\usage{
ee_composite(x, ...)

\method{ee_composite}{tidyee}(x, stat, ...)
}
\arguments{
\item{x}{tidyee object containing \code{ee$ImageCollection}}

\item{...}{other arguments}

\item{stat}{A \code{character} indicating what to reduce the ImageCollection by,
e.g. 'median' (default), 'mean',  'max', 'min', 'sum', 'sd', 'first'.}
}
\value{
tidyee class containing \code{ee$Image} where all images within \code{ee$ImageCollection} have been aggregated based on pixel-level stats
}
\description{
ee_composite
}
