% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_temporal_composites.R
\name{ee_month_composite}
\alias{ee_month_composite}
\title{Pixel-level composite by month}
\usage{
ee_month_composite(imageCol, stat, month, ...)
}
\arguments{
\item{imageCol}{An earth engine ImageCollection}

\item{...}{extra args to pass on}
}
\description{
Pixel-level composite by month
}
