% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_year_composite.R
\name{ee_year_composite}
\alias{ee_year_composite}
\title{Filter by Year}
\usage{
ee_year_composite(imageCol, stat, startDate, endDate, ...)
}
\arguments{
\item{imageCol}{An earth engine ImageCollection}

\item{stat}{A \code{character} indicating what to reduce the imageCollection by,
e.g. 'median' (default), 'mean',  'max', 'min', 'sum', 'sd', 'first'.}

\item{startDate}{\code{character} format date, e.g. "2018-10-23".}

\item{endDate}{\code{character} format date, e.g. "2018-10-23".}

\item{...}{extra args to pass on}
}
\description{
Filter by Year
}
