% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_temporal_composites.R
\name{ee_year_composite}
\alias{ee_year_composite}
\title{Pixel level composite by year}
\usage{
ee_year_composite(imageCol, stat, year, ...)
}
\arguments{
\item{imageCol}{ee$ImageCollection}

\item{stat}{A \code{character} indicating what to reduce the imageCollection by,
e.g. 'median' (default), 'mean',  'max', 'min', 'sum', 'sd', 'first'.}

\item{year}{\code{numeric} vector containing years (i.e c(2001,2002,2003))}

\item{...}{other arguments}
}
\description{
Pixel level composite by year
}
