% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_temporal_filters.R
\name{ee_year_filter}
\alias{ee_year_filter}
\title{ee_year_filter}
\usage{
ee_year_filter(imageCol, year, ...)
}
\arguments{
\item{imageCol}{ee$ImageCollection}

\item{year}{\code{numeric} vector containing years (i.e c(2001,2002,2003))}

\item{...}{other arguments}
}
\value{
ee$ImageCollection or ee$Image filtered by year
}
\description{
ee_year_filter
}
