% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_by}
\alias{group_by}
\title{Group an imageCollection or tidyee object with Imagecollections by a parameter}
\arguments{
\item{x}{ee$ImageCollection or tidyee object}

\item{...}{group_by variables}
}
\value{
ee$ImageCollection with grouped_vars attribute
}
\description{
Group an imageCollection or tidyee object with Imagecollections by a parameter
}
\examples{
\dontrun{
library(tidyrgee)
ee_Initialize()
modis_ic <- ee$ImageCollection("MODIS/006/MOD13Q1")
modis_ic |>
   filter(date>="2016-01-01",date<="2019-12-31") |>
   group_by(year)
}
}
\seealso{
\code{\link[dplyr]{group_by}} for information about group_by on normal data tables.
}
