% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_join.R
\name{inner_join}
\alias{inner_join}
\title{inner_join bands from different image/ImageCollections based on shared property}
\arguments{
\item{x, y}{A     pair of tidyee objects containing ee$ImageCollections}

\item{by}{A character vector of variables to join by.}
}
\value{
An object of the same type as \code{x}. The output has the following properties:
Same number of images as \code{x}
Total number of bands equal the number of bands in \code{x} plus the number of bands in \code{y}
}
\description{
inner_join bands from different image/ImageCollections based on shared property
}
\seealso{
\code{\link[dplyr]{inner_join}} for information about inner_join on normal data tables.
}
