% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-tidyee.R
\name{mutate}
\alias{mutate}
\title{mutate}
\usage{
mutate(x, ...)
}
\arguments{
\item{x}{tidyee class object (list of ee_ob, vrt)}

\item{...}{mutate arguments}
}
\value{
tidyee object with vrt component containing new mutated cols
}
\description{
mutate
}
\examples{
\dontrun{
library(tidyrgee)
library(rgee)
ee_Initialize()
modis_ic <- ee$ImageCollection("MODIS/006/MOD13Q1")
modis_ic_tidy <- as_tidyee(modis_ic)
}
}
