% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str}
\alias{str}
\title{Compactly Display the Structure of an Arbitrary R Object}
\arguments{
\item{object}{imageCollection or tidyee class object}

\item{...}{potential further arguments (required for Method/Generic reasons).}
}
\value{
return str
}
\description{
Compactly Display the Structure of an Arbitrary R Object
}
