% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str}
\alias{str}
\title{Compactly Display the Structure of an Arbitrary R Object}
\arguments{
\item{x}{imageCollection or tidyee class object}
}
\value{
return str
}
\description{
Compactly Display the Structure of an Arbitrary R Object
}
