% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise}
\alias{summarise}
\title{summarise (method for imageCollections)}
\usage{
summarise(x, stat, ...)
}
\arguments{
\item{x}{ee$Image or ee$ImageCollection}

\item{stat}{stat/function to apply}

\item{...}{other arguments}
}
\value{
ee$Image or ee$ImageCollection where pixels are summarised by group_by and stat
}
\description{
summarise (method for imageCollections)
}
\examples{
\dontrun{
library(tidyrgee)
ee_Initialize()
modis_ic <- ee$ImageCollection("MODIS/006/MOD13Q1")
modis_ic |>
   filter(date>="2016-01-01",date<="2019-12-31") |>
   group_by(year) |>
   summarise(stat="max")
}

}
